moving the test from the controller to the ui service
